+++ CHANGELOG.md
@@ -219,3 +219,4 @@
 - 2023-12-21: Init mod, main, LICENSE, qualify, +3
 - 2023-12-22: Test Makefile, lib, mod, qualify, +6
 - 2023-12-22: Integ conf, frontier, chat, discover, +5
+- 2023-12-23: Health leads, geo, frontier, lib, +5